extracted the massive process_agent_response method (~325 lines) to agent_response_handler.rs
